fix: update outline sidebar hooks for plugins
