+++ src/courseware/course/sidebar/sidebars/course-outline/components/SidebarSection.jsx
@@ -1,13 +1,12 @@
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { useSelector } from 'react-redux';
 import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 import { Button, Icon } from '@openedx/paragon';
 import { ChevronRight as ChevronRightIcon } from '@openedx/paragon/icons';
 
 import courseOutlineMessages from '@src/course-home/outline-tab/messages';
-import { getSequenceId } from '@src/courseware/data/selectors';
 import CompletionIcon from './CompletionIcon';
+import { useCourseOutlineSidebar } from '../hooks';
 
 const SidebarSection = ({ intl, section, handleSelectSection }) => {
   const {
@@ -18,7 +17,7 @@ const SidebarSection = ({ intl, section, handleSelectSection }) => {
     completionStat,
   } = section;
 
-  const activeSequenceId = useSelector(getSequenceId);
+  const { activeSequenceId } = useCourseOutlineSidebar();
   const isActiveSection = sequenceIds.includes(activeSequenceId);
 
   const sectionTitle = (
