+++ src/courseware/course/sidebar/sidebars/course-outline/components/SidebarSection.test.jsx
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { IntlProvider } from '@edx/frontend-platform/i18n';
@@ -5,6 +6,7 @@ import { AppProvider } from '@edx/frontend-platform/react';
 
 import { initializeTestStore } from '@src/setupTest';
 import courseOutlineMessages from '@src/course-home/outline-tab/messages';
+import SidebarContext from '../../../SidebarContext';
 import SidebarSection from './SidebarSection';
 
 describe('<SidebarSection />', () => {
@@ -19,17 +21,23 @@ describe('<SidebarSection />', () => {
     section = state.courseware.courseOutline.sections[activeSectionId];
   };
 
-  const RootWrapper = (props) => (
-    <AppProvider store={store} wrapWithRouter={false}>
-      <IntlProvider locale="en">
-        <SidebarSection
-          section={section}
-          handleSelectSection={mockHandleSelectSection}
-          {...props}
-        />,
-      </IntlProvider>
-    </AppProvider>
-  );
+  const RootWrapper = (props) => {
+    const mockData = useMemo(() => ({ toggleSidebar: jest.fn() }), []);
+
+    return (
+      <AppProvider store={store} wrapWithRouter={false}>
+        <IntlProvider locale="en">
+          <SidebarContext.Provider value={mockData}>
+            <SidebarSection
+              section={section}
+              handleSelectSection={mockHandleSelectSection}
+              {...props}
+            />
+          </SidebarContext.Provider>
+        </IntlProvider>
+      </AppProvider>
+    );
+  };
 
   beforeEach(() => {
     mockHandleSelectSection = jest.fn();
