+++ src/courseware/course/sidebar/sidebars/course-outline/components/SidebarUnit.jsx
@@ -1,14 +1,10 @@
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { Link } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
 import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
-import { sendTrackEvent, sendTrackingLogEvent } from '@edx/frontend-platform/analytics';
 
-import { checkBlockCompletion } from '@src/courseware/data';
-import { getCourseOutline } from '@src/courseware/data/selectors';
 import messages from '../messages';
 import UnitIcon, { UNIT_ICON_TYPES } from './UnitIcon';
+import UnitLinkWrapper from './UnitLinkWrapper';
 
 const SidebarUnit = ({
   id,
@@ -26,43 +22,18 @@ const SidebarUnit = ({
     title,
     icon = UNIT_ICON_TYPES.other,
   } = unit;
-  const dispatch = useDispatch();
-  const { sequences = {} } = useSelector(getCourseOutline);
-
-  const logEvent = (eventName, widgetPlacement) => {
-    const findSequenceByUnitId = (unitId) => Object.values(sequences).find(seq => seq.unitIds.includes(unitId));
-    const activeSequence = findSequenceByUnitId(activeUnitId);
-    const targetSequence = findSequenceByUnitId(id);
-    const payload = {
-      id: activeUnitId,
-      current_tab: activeSequence.unitIds.indexOf(activeUnitId) + 1,
-      tab_count: activeSequence.unitIds.length,
-      target_id: id,
-      target_tab: targetSequence.unitIds.indexOf(id) + 1,
-      widget_placement: widgetPlacement,
-    };
-
-    if (activeSequence.id !== targetSequence.id) {
-      payload.target_tab_count = targetSequence.unitIds.length;
-    }
-
-    sendTrackEvent(eventName, payload);
-    sendTrackingLogEvent(eventName, payload);
-  };
-
-  const handleClick = () => {
-    logEvent('edx.ui.lms.sequence.tab_selected', 'left');
-    dispatch(checkBlockCompletion(courseId, sequenceId, activeUnitId));
-  };
 
   const iconType = isLocked ? UNIT_ICON_TYPES.lock : icon;
 
   return (
     <li className={classNames({ 'bg-info-100': isActive, 'border-top border-light': !isFirst })}>
-      <Link
-        to={`/course/${courseId}/${sequenceId}/${id}`}
-        className="row w-100 m-0 d-flex align-items-center text-gray-700"
-        onClick={handleClick}
+      <UnitLinkWrapper
+        {...{
+          sequenceId,
+          activeUnitId,
+          id,
+          courseId,
+        }}
       >
         <div className="col-auto p-0">
           <UnitIcon type={iconType} isCompleted={complete} />
@@ -75,7 +46,7 @@ const SidebarUnit = ({
             , {intl.formatMessage(complete ? messages.completedUnit : messages.incompleteUnit)}
           </span>
         </div>
-      </Link>
+      </UnitLinkWrapper>
     </li>
   );
 };
